Write a lot of bridges
