+++ wfc_core/src/bitvec.rs
@@ -1,29 +1,31 @@
 use std::fmt;
 
-const DATA_MASK: u64 = u64::MAX >> 8;
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct BitVec {
+pub struct BitVec<const N: usize> {
     /// The data and length of the bit vector.
     ///
-    /// The last eight bits represent the length and the first 248 bits contain
-    /// the data itself. While the length could be computed every time with
-    /// [`u64::count_ones`], the intrinsic showed up in profiling quite a lot
-    /// and it is just faster to use the cached value.
-    data: [u64; 4],
+    /// XXX: Doc where is data and where is len.
+    data: [u64; N],
 }
 
-impl BitVec {
-    pub const MAX_LEN: u8 = u8::MAX - 8;
+impl<const N: usize> BitVec<N> {
+    // TODO(yan): @Correctness This implementation completely breaks down once
+    // DATA_CAP reaches 2^64 (or N reaches 2^60), because len currently can't span
+    // multiple blocks. In practice, this is a lot of memory and won't happen.
+
+    pub const DATA_CAP: usize = N * u64::BITS as usize - Self::LEN_SEGMENT_SIZE;
+
+    const LEN_SEGMENT_SIZE: usize = find_len_segment_size(N);
+    const DATA_MASK: u64 = u64::MAX >> (64 - Self::LEN_SEGMENT_SIZE);
 
     /// Creates a new bit vector with all bits set to zero.
     pub const fn zeros() -> Self {
-        Self { data: [0; 4] }
+        Self { data: [0; N] }
     }
 
     /// Returns whether the bit vector has a bit set.
-    pub fn contains(&self, index: u8) -> bool {
-        assert!(index < Self::MAX_LEN);
+    pub fn contains(&self, index: usize) -> bool {
+        assert!(index < Self::DATA_CAP);
         let index = usize::from(index);
 
         let blk_index = index / 64;
@@ -35,8 +37,8 @@ impl BitVec {
 
     /// Sets a bit in the vector to one. Returns [`true`] if the bit was not
     /// previously set.
-    pub fn add(&mut self, index: u8) -> bool {
-        assert!(index < Self::MAX_LEN);
+    pub fn add(&mut self, index: usize) -> bool {
+        assert!(index < Self::DATA_CAP);
         let index = usize::from(index);
 
         let blk_index = index / 64;
@@ -55,8 +57,8 @@ impl BitVec {
 
     /// Sets a bit in the vector to zero. Returns [`true`] if the bit was
     /// previously set.
-    pub fn remove(&mut self, index: u8) -> bool {
-        assert!(index < Self::MAX_LEN);
+    pub fn remove(&mut self, index: usize) -> bool {
+        assert!(index < Self::DATA_CAP);
         let index = usize::from(index);
 
         let blk_index = index / 64;
@@ -76,16 +78,15 @@ impl BitVec {
     /// Returns the number of ones in the binary representation of the bit
     /// vector.
     pub fn len(&self) -> usize {
-        (self.data[3] >> 56) as usize
+        (self.data[N - 1] >> Self::LEN_SEGMENT_SIZE) as usize
     }
 
     /// Sets all bits in the bit vector to zero.
     pub fn clear(&mut self) {
-        self.data = [0; 4];
+        self.data = [0; N];
     }
 
-    /// Creates a shared borrowed iterator over the indices of bits set to one.
-    pub fn iter(&self) -> BitVecIterator {
+    pub fn iter(&self) -> BitVecIterator<N> {
         BitVecIterator {
             bitvec: self,
             next: 0,
@@ -93,47 +94,61 @@ impl BitVec {
     }
 
     fn inc_len(&mut self) {
-        let mut len = self.data[3] >> 56;
-        assert!(len < Self::MAX_LEN as u64);
+        let mut len = self.data[N - 1] >> Self::LEN_SEGMENT_SIZE;
+        assert!(len < Self::DATA_CAP as u64);
 
         len += 1;
 
-        self.data[3] = (self.data[3] & DATA_MASK) | (len << 56);
+        self.data[N - 1] = (self.data[N - 1] & Self::DATA_MASK) | (len << Self::LEN_SEGMENT_SIZE);
     }
 
     fn dec_len(&mut self) {
-        let mut len = self.data[3] >> 56;
+        let mut len = self.data[N - 1] >> Self::LEN_SEGMENT_SIZE;
         assert!(len > 0);
 
         len -= 1;
 
-        self.data[3] = (self.data[3] & DATA_MASK) | (len << 56)
+        self.data[N - 1] = (self.data[N - 1] & Self::DATA_MASK) | (len << Self::LEN_SEGMENT_SIZE)
     }
 }
 
-impl fmt::Binary for BitVec {
+impl<const N: usize> fmt::Binary for BitVec<N> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "[{:064b}][{:064b}][{:064b}][{:064b}]",
-            self.data[0], self.data[1], self.data[2], self.data[3],
-        )
+        write!(f, "B:{}", N)?;
+
+        for i in 0..N {
+            write!(f, "[{:064b}]", self.data[i])?;
+        }
+
+        Ok(())
     }
 }
 
-pub struct BitVecIterator<'a> {
-    bitvec: &'a BitVec,
-    next: u8,
+impl<'a, const N: usize> IntoIterator for &'a BitVec<N> {
+    type Item = usize;
+    type IntoIter = BitVecIterator<'a, N>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        BitVecIterator {
+            bitvec: self,
+            next: 0,
+        }
+    }
 }
 
-impl<'a> Iterator for BitVecIterator<'a> {
-    type Item = u8;
+pub struct BitVecIterator<'a, const N: usize> {
+    bitvec: &'a BitVec<N>,
+    next: usize,
+}
 
-    fn next(&mut self) -> Option<u8> {
+impl<'a, const N: usize> Iterator for BitVecIterator<'a, N> {
+    type Item = usize;
+
+    fn next(&mut self) -> Option<usize> {
         // TODO(yan): @Speed Try using leading zeros intrinsic to directly seek
         // to the next element. Ideally do this in a BitVec::first_after
         // function or similar.
-        while self.next < BitVec::MAX_LEN {
+        while self.next < BitVec::<N>::DATA_CAP {
             let index = self.next;
 
             let contains = self.bitvec.contains(index);
@@ -148,46 +163,49 @@ impl<'a> Iterator for BitVecIterator<'a> {
     }
 
     fn size_hint(&self) -> (usize, Option<usize>) {
+        // XXX: Can we cast less? If we could be u16, we could widen to u32 and
+        // usize as needed without fear.
+
         // Size hints are computed using the count_ones intrinsic on the u64
         // components of the bit vector. To not count bits we already iterated
         // over, we mask them. Masks are computed from the the index to the bit
         // we would visit next.
 
-        let n = u32::from(self.next);
+        // TODO(yan): @Correctness self.next likely won't be larger than than
+        // u32, but we'd love to prove this statically. Unfortunately, we need
+        // to bound N with usize, because it's a parameter for array size.
+        let next = self.next as u32;
 
-        let shl0 = n;
-        let shl1 = n.saturating_sub(64);
-        let shl2 = n.saturating_sub(64 * 2);
-        let shl3 = n.saturating_sub(64 * 3);
+        let mut size = 0;
+        for i in 0..N - 1 {
+            let s = next.saturating_sub(64 * i as u32);
+            let m = u64::MAX.checked_shl(s).unwrap_or(0);
+            let c = (m & self.bitvec.data[i]).count_ones();
 
-        let m0 = u64::MAX.checked_shl(shl0).unwrap_or(0);
-        let m1 = u64::MAX.checked_shl(shl1).unwrap_or(0);
-        let m2 = u64::MAX.checked_shl(shl2).unwrap_or(0);
-        let m3 = u64::MAX.checked_shl(shl3).unwrap_or(0) & DATA_MASK;
+            size += c;
+        }
 
-        // Mask out spent bits and count how many set bits remain.
+        {
+            let s = next.saturating_sub(64 * (N as u32 - 1));
+            let m = u64::MAX.checked_shl(s).unwrap_or(0) & BitVec::<N>::DATA_MASK;
+            let c = (m & self.bitvec.data[N - 1]).count_ones();
 
-        // usize is defined to be at least 16 bits wide, the following `as`
-        // casts should be ok for up to 2^16 ones in the whole array.
-        let c0 = (m0 & self.bitvec.data[0]).count_ones() as usize;
-        let c1 = (m1 & self.bitvec.data[1]).count_ones() as usize;
-        let c2 = (m2 & self.bitvec.data[2]).count_ones() as usize;
-        let c3 = (m3 & self.bitvec.data[3]).count_ones() as usize;
+            size += c;
+        }
 
-        let size = c0 + c1 + c2 + c3;
-
-        (size, Some(size))
+        (size as usize, Some(size as usize))
     }
 }
 
-impl<'a> IntoIterator for &'a BitVec {
-    type Item = u8;
-    type IntoIter = BitVecIterator<'a>;
+const fn find_len_segment_size(blk_count: usize) -> usize {
+    // TODO(yan): This is kind of lazy, work out the formula?
 
-    fn into_iter(self) -> Self::IntoIter {
-        BitVecIterator {
-            bitvec: self,
-            next: 0,
-        }
+    // Start at 6, as 2^6 is 64 - our block size.
+    let mut offset = 6;
+
+    while blk_count * 64 > 2_usize.pow(offset as u32) {
+        offset += 1;
     }
+
+    offset
 }
