--- wfc_core/src/slots.rs
@@ -1,297 +1,0 @@
-use std::fmt;
-
-// pub struct Slots {
-//     len: usize,
-//     data: SlotsData,
-// }
-
-// #[derive(Clone)]
-// enum SlotsData {
-//     // One block, 64 bits, 6-bit addressing, 58 data bits
-//     Size1(Vec<BitVec<1>>),
-//     // Two blocks - 128 bits, 7-bit addressing, 121 data bits
-//     Size2(Vec<BitVec<2>>),
-//     // Four blocks - 256 bits, 8-bit addressing, 248 data bits
-//     Size4(Vec<BitVec<4>>),
-//     // Eight blocks - 512 bits, 9-bit addressing, 503 data bits
-//     Size8(Vec<BitVec<8>>),
-//     // Sixteen blocks - 1024 bits, 10-bit addressing, 1014 data bits
-//     Size16(Vec<BitVec<16>>),
-// }
-
-// impl Slots {
-//     pub const MAX_MODULE_COUNT: usize = 1014;
-
-//     pub fn new(slot_count: usize, module_count: usize) -> Option<Slots> {
-//         let data = match module_count {
-//             0..=58 => Some(SlotsData::Size1(vec![BitVec::zeros(); slot_count])),
-//             59..=121 => Some(SlotsData::Size2(vec![BitVec::zeros(); slot_count])),
-//             122..=248 => Some(SlotsData::Size4(vec![BitVec::zeros(); slot_count])),
-//             249..=503 => Some(SlotsData::Size8(vec![BitVec::zeros(); slot_count])),
-//             504..=1014 => Some(SlotsData::Size16(vec![BitVec::zeros(); slot_count])),
-//             _ => None,
-//         };
-
-//         data.map(|data| Slots { len: 0, data })
-//     }
-
-//     pub fn contains(&self, slot_index: usize, module_index: usize) -> bool {
-//         match self.data {
-//             SlotsData::Size1(slots) => slots[slot_index].contains(module_index),
-//             SlotsData::Size2(slots) => slots[slot_index].contains(module_index),
-//             SlotsData::Size4(slots) => slots[slot_index].contains(module_index),
-//             SlotsData::Size8(slots) => slots[slot_index].contains(module_index),
-//             SlotsData::Size16(slots) => slots[slot_index].contains(module_index),
-//         }
-//     }
-
-//     pub fn add(&mut self, slot_index: usize, module_index: usize) -> bool {
-//         match self.data {
-//             SlotsData::Size1(slots) => slots[slot_index].add(module_index),
-//             SlotsData::Size2(slots) => slots[slot_index].add(module_index),
-//             SlotsData::Size4(slots) => slots[slot_index].add(module_index),
-//             SlotsData::Size8(slots) => slots[slot_index].add(module_index),
-//             SlotsData::Size16(slots) => slots[slot_index].add(module_index),
-//         }
-//     }
-
-//     pub fn remove(&mut self, slot_index: usize, module_index: usize) -> bool {
-//         match self.data {
-//             SlotsData::Size1(slots) => slots[slot_index].remove(module_index),
-//             SlotsData::Size2(slots) => slots[slot_index].remove(module_index),
-//             SlotsData::Size4(slots) => slots[slot_index].remove(module_index),
-//             SlotsData::Size8(slots) => slots[slot_index].remove(module_index),
-//             SlotsData::Size16(slots) => slots[slot_index].remove(module_index),
-//         }
-//     }
-
-//     pub fn slot_count(&self) -> usize {
-//         match self.data {
-//             SlotsData::Size1(slots) => slots.len(),
-//             SlotsData::Size2(slots) => slots.len(),
-//             SlotsData::Size4(slots) => slots.len(),
-//             SlotsData::Size8(slots) => slots.len(),
-//             SlotsData::Size16(slots) => slots.len(),
-//         }
-//     }
-
-//     pub fn module_capacity(&self) -> usize {
-//         match self.data {
-//             SlotsData::Size1(slots) => 58,
-//             SlotsData::Size2(slots) => 121,
-//             SlotsData::Size4(slots) => 248,
-//             SlotsData::Size8(slots) => 503,
-//             SlotsData::Size16(slots) => 1014,
-//         }
-//     }
-// }
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct BitVec<const N: usize> {
-    /// The data and length of the bit vector.
-    ///
-    /// XXX: Doc where is data and where is len.
-    data: [u64; N],
-}
-
-impl<const N: usize> BitVec<N> {
-    // TODO(yan): @Correctness This implementation completely breaks down once
-    // DATA_CAP reaches 2^64 (or N reaches 2^60), because len currently can't span
-    // multiple blocks. In practice, this is a lot of memory and won't happen.
-
-    pub const DATA_CAP: usize = N * u64::BITS as usize - Self::LEN_SEGMENT_SIZE;
-
-    const LEN_SEGMENT_SIZE: usize = find_len_segment_size(N);
-    const DATA_MASK: u64 = u64::MAX >> (64 - Self::LEN_SEGMENT_SIZE);
-
-    /// Creates a new bit vector with all bits set to zero.
-    pub const fn zeros() -> Self {
-        Self { data: [0; N] }
-    }
-
-    /// Returns whether the bit vector has a bit set.
-    pub fn contains(&self, index: usize) -> bool {
-        assert!(index < Self::DATA_CAP);
-        let index = usize::from(index);
-
-        let blk_index = index / 64;
-        let bit_index = index % 64;
-        let value = self.data[blk_index] & (1 << bit_index);
-
-        value != 0
-    }
-
-    /// Sets a bit in the vector to one. Returns [`true`] if the bit was not
-    /// previously set.
-    pub fn add(&mut self, index: usize) -> bool {
-        assert!(index < Self::DATA_CAP);
-        let index = usize::from(index);
-
-        let blk_index = index / 64;
-        let bit_index = index % 64;
-        let value = self.data[blk_index] & (1 << bit_index);
-
-        self.data[blk_index] |= 1 << bit_index;
-
-        if value == 0 {
-            self.inc_len();
-            true
-        } else {
-            false
-        }
-    }
-
-    /// Sets a bit in the vector to zero. Returns [`true`] if the bit was
-    /// previously set.
-    pub fn remove(&mut self, index: usize) -> bool {
-        assert!(index < Self::DATA_CAP);
-        let index = usize::from(index);
-
-        let blk_index = index / 64;
-        let bit_index = index % 64;
-        let value = self.data[blk_index] & (1 << bit_index);
-
-        self.data[blk_index] &= !(1 << bit_index);
-
-        if value != 0 {
-            self.dec_len();
-            true
-        } else {
-            false
-        }
-    }
-
-    /// Returns the number of ones in the binary representation of the bit
-    /// vector.
-    pub fn len(&self) -> usize {
-        (self.data[N - 1] >> Self::LEN_SEGMENT_SIZE) as usize
-    }
-
-    /// Sets all bits in the bit vector to zero.
-    pub fn clear(&mut self) {
-        self.data = [0; N];
-    }
-
-    pub fn iter(&self) -> BitVecIterator<N> {
-        BitVecIterator {
-            bitvec: self,
-            next: 0,
-        }
-    }
-
-    fn inc_len(&mut self) {
-        let mut len = self.data[N - 1] >> Self::LEN_SEGMENT_SIZE;
-        assert!(len < Self::DATA_CAP as u64);
-
-        len += 1;
-
-        self.data[N - 1] = (self.data[N - 1] & Self::DATA_MASK) | (len << Self::LEN_SEGMENT_SIZE);
-    }
-
-    fn dec_len(&mut self) {
-        let mut len = self.data[N - 1] >> Self::LEN_SEGMENT_SIZE;
-        assert!(len > 0);
-
-        len -= 1;
-
-        self.data[N - 1] = (self.data[N - 1] & Self::DATA_MASK) | (len << Self::LEN_SEGMENT_SIZE)
-    }
-}
-
-impl<const N: usize> fmt::Binary for BitVec<N> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "B:{}", N)?;
-
-        for i in 0..N {
-            write!(f, "[{:064b}]", self.data[i])?;
-        }
-
-        Ok(())
-    }
-}
-
-impl<'a, const N: usize> IntoIterator for &'a BitVec<N> {
-    type Item = usize;
-    type IntoIter = BitVecIterator<'a, N>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        BitVecIterator {
-            bitvec: self,
-            next: 0,
-        }
-    }
-}
-
-pub struct BitVecIterator<'a, const N: usize> {
-    bitvec: &'a BitVec<N>,
-    next: usize,
-}
-
-impl<'a, const N: usize> Iterator for BitVecIterator<'a, N> {
-    type Item = usize;
-
-    fn next(&mut self) -> Option<usize> {
-        // TODO(yan): @Speed Try using leading zeros intrinsic to directly seek
-        // to the next element. Ideally do this in a BitVec::first_after
-        // function or similar.
-        while self.next < BitVec::<N>::DATA_CAP {
-            let index = self.next;
-
-            let contains = self.bitvec.contains(index);
-            self.next += 1;
-
-            if contains {
-                return Some(index);
-            }
-        }
-
-        None
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        // XXX: Can we cast less? If we could be u16, we could widen to u32 and
-        // usize as needed without fear.
-
-        // Size hints are computed using the count_ones intrinsic on the u64
-        // components of the bit vector. To not count bits we already iterated
-        // over, we mask them. Masks are computed from the the index to the bit
-        // we would visit next.
-
-        // TODO(yan): @Correctness self.next likely won't be larger than than
-        // u32, but we'd love to prove this statically. Unfortunately, we need
-        // to bound N with usize, because it's a parameter for array size.
-        let next = self.next as u32;
-
-        let mut size = 0;
-        for i in 0..N-1 {
-            let s = next.saturating_sub(64 * i as u32);
-            let m = u64::MAX.checked_shl(s).unwrap_or(0);
-            let c = (m & self.bitvec.data[i]).count_ones();
-
-            size += c;
-        }
-
-        {
-            let s = next.saturating_sub(64 * (N as u32 - 1));
-            let m = u64::MAX.checked_shl(s).unwrap_or(0) & BitVec::<N>::DATA_MASK;
-            let c = (m & self.bitvec.data[N - 1]).count_ones();
-
-            size += c;
-        }
-
-        (size as usize, Some(size as usize))
-    }
-}
-
-const fn find_len_segment_size(blk_count: usize) -> usize {
-    // TODO(yan): This is kind of lazy, work out the formula?
-
-    // Start at 6, as 2^6 is 64 - our block size.
-    let mut offset = 6;
-
-    while blk_count * 64 > 2_usize.pow(offset as u32) {
-        offset += 1;
-    }
-
-    offset
-}
